Optimize test structure for previously created tests
